consolidate dataframe_subquery.rs into dataframe.rs
